+++ splitio/producer/initialization.go
@@ -3,20 +3,14 @@ package producer
 import (
 	"errors"
 	"fmt"
-	"log"
-	"net/url"
 	"time"
 
 	cconf "github.com/splitio/go-split-commons/v4/conf"
 	"github.com/splitio/go-split-commons/v4/dtos"
-	"github.com/splitio/go-split-commons/v4/provisional"
-	"github.com/splitio/go-split-commons/v4/provisional/strategy"
 	"github.com/splitio/go-split-commons/v4/service/api"
-	storageCommon "github.com/splitio/go-split-commons/v4/storage"
 	"github.com/splitio/go-split-commons/v4/storage/inmemory"
 	"github.com/splitio/go-split-commons/v4/storage/redis"
 	"github.com/splitio/go-split-commons/v4/synchronizer"
-	"github.com/splitio/go-split-commons/v4/synchronizer/worker/impressionscount"
 	"github.com/splitio/go-split-commons/v4/synchronizer/worker/segment"
 	"github.com/splitio/go-split-commons/v4/synchronizer/worker/split"
 	"github.com/splitio/go-split-commons/v4/tasks"
@@ -34,9 +28,7 @@ import (
 	"github.com/splitio/split-synchronizer/v5/splitio/producer/task"
 	"github.com/splitio/split-synchronizer/v5/splitio/producer/worker"
 	hcApplication "github.com/splitio/split-synchronizer/v5/splitio/provisional/healthcheck/application"
-	hcAppCounter "github.com/splitio/split-synchronizer/v5/splitio/provisional/healthcheck/application/counter"
 	hcServices "github.com/splitio/split-synchronizer/v5/splitio/provisional/healthcheck/services"
-	hcServicesCounter "github.com/splitio/split-synchronizer/v5/splitio/provisional/healthcheck/services/counter"
 	"github.com/splitio/split-synchronizer/v5/splitio/provisional/observability"
 	"github.com/splitio/split-synchronizer/v5/splitio/util"
 )
@@ -284,82 +276,4 @@ func Start(logger logging.LoggerInterface, cfg *conf.Main) error {
 	rtm.RegisterShutdownHandler()
 	rtm.Block()
 	return nil
-}
-
-func getAppCounterConfigs(storage storageCommon.SplitStorage) (hcAppCounter.ThresholdConfig, hcAppCounter.ThresholdConfig, hcAppCounter.PeriodicConfig) {
-	splitsConfig := hcAppCounter.DefaultThresholdConfig("Splits")
-	segmentsConfig := hcAppCounter.DefaultThresholdConfig("Segments")
-	storageConfig := hcAppCounter.PeriodicConfig{
-		Name:                     "Storage",
-		MaxErrorsAllowedInPeriod: 5,
-		Period:                   3600,
-		Severity:                 hcAppCounter.Low,
-		ValidationFunc: func(c hcAppCounter.PeriodicCounterInterface) {
-			_, err := storage.ChangeNumber()
-			if err != nil {
-				c.NotifyError()
-			}
-		},
-		ValidationFuncPeriod: 10,
-	}
-
-	return splitsConfig, segmentsConfig, storageConfig
-}
-
-func getServicesCountersConfig(advanced *cconf.AdvancedConfig) []hcServicesCounter.Config {
-	var cfgs []hcServicesCounter.Config
-
-	apiConfig := hcServicesCounter.DefaultConfig("API", advanced.SdkURL, "/version")
-	eventsConfig := hcServicesCounter.DefaultConfig("Events", advanced.EventsURL, "/version")
-	authConfig := hcServicesCounter.DefaultConfig("Auth", advanced.AuthServiceURL, "/health")
-
-	telemetryURL, err := url.Parse(advanced.TelemetryServiceURL)
-	if err != nil {
-		log.Fatal(err)
-	}
-	telemetryConfig := hcServicesCounter.DefaultConfig("Telemetry", fmt.Sprintf("%s://%s", telemetryURL.Scheme, telemetryURL.Host), "/health")
-
-	streamingURL, err := url.Parse(advanced.StreamingServiceURL)
-	if err != nil {
-		log.Fatal(err)
-	}
-	streamingConfig := hcServicesCounter.DefaultConfig("Streaming", fmt.Sprintf("%s://%s", streamingURL.Scheme, streamingURL.Host), "/health")
-
-	return append(cfgs, telemetryConfig, authConfig, apiConfig, eventsConfig, streamingConfig)
-}
-
-func buildImpressionManager(
-	impressionsMode string,
-	impListener impressionlistener.ImpressionBulkListener,
-	runtimeTelemetry storageCommon.TelemetryRuntimeProducer,
-	splitTasks *synchronizer.SplitTasks,
-	workers *synchronizer.Workers,
-	splitAPI *api.SplitAPI,
-	metadata dtos.Metadata,
-	logger logging.LoggerInterface,
-) (provisional.ImpressionManager, error) {
-	listenerEnabled := impListener != nil
-	switch impressionsMode {
-	case cconf.ImpressionsModeDebug:
-		impressionObserver, err := strategy.NewImpressionObserver(500)
-		if err != nil {
-			return nil, err
-		}
-
-		strategy := strategy.NewDebugImpl(impressionObserver, listenerEnabled)
-
-		return provisional.NewImpressionManager(strategy), nil
-	default:
-		impressionsCounter := strategy.NewImpressionsCounter()
-		impressionObserver, err := strategy.NewImpressionObserver(500)
-		if err != nil {
-			return nil, err
-		}
-
-		workers.ImpressionsCountRecorder = impressionscount.NewRecorderSingle(impressionsCounter, splitAPI.ImpressionRecorder, metadata, logger, runtimeTelemetry)
-		splitTasks.ImpressionsCountSyncTask = tasks.NewRecordImpressionsCountTask(workers.ImpressionsCountRecorder, logger, 500) // TODO: update period
-		strategy := strategy.NewOptimizedImpl(impressionObserver, impressionsCounter, runtimeTelemetry, listenerEnabled)
-
-		return provisional.NewImpressionManager(strategy), nil
-	}
 }
